Restructure Psi4
